backend/requirements.txt
flask
torch
torchvision
numpy
Pillow